support for weekly stats starting on mondays #16
