+++ redis_metrics/tests/test_settings.py
@@ -12,6 +12,7 @@ from .. import settings as rs
 @override_settings(REDIS_METRICS_SOCKET_CONNECTION_POOL=None)
 @override_settings(REDIS_METRICS_MIN_GRANULARITY='seconds')
 @override_settings(REDIS_METRICS_MAX_GRANULARITY='yearly')
+@override_settings(REDIS_METRICS_MONDAY_FIRST_DAY_OF_WEEK=False)
 class TestAppSettings(TestCase):
 
     def test_settings_getattr(self):
@@ -23,6 +24,7 @@ class TestAppSettings(TestCase):
         self.assertEquals(rs.app_settings.REDIS_METRICS_SOCKET_CONNECTION_POOL, None)
         self.assertEquals(rs.app_settings.REDIS_METRICS_MIN_GRANULARITY, 'seconds')
         self.assertEquals(rs.app_settings.REDIS_METRICS_MAX_GRANULARITY, 'yearly')
+        self.assertEquals(rs.app_settings.REDIS_METRICS_MONDAY_FIRST_DAY_OF_WEEK, False)
 
     def test_settings_getitem(self):
         self.assertEquals(rs.app_settings['REDIS_METRICS_HOST'], 'localhost')
@@ -33,17 +35,11 @@ class TestAppSettings(TestCase):
         self.assertEquals(rs.app_settings['REDIS_METRICS_SOCKET_CONNECTION_POOL'], None)
         self.assertEquals(rs.app_settings['REDIS_METRICS_MIN_GRANULARITY'], 'seconds')
         self.assertEquals(rs.app_settings['REDIS_METRICS_MAX_GRANULARITY'], 'yearly')
+        self.assertEquals(rs.app_settings['REDIS_METRICS_MONDAY_FIRST_DAY_OF_WEEK'], False)
 
     def test_settings_getattr_raises_attribute_error(self):
         with self.assertRaises(AttributeError):
             rs.app_settings.REDIS_METRICS_LOLWUT
-
-    def test_metric_key_patterns(self):
-        # These two things should always be the same.
-        self.assertEqual(
-            sorted(rs.METRIC_KEY_PATTERNS.keys()),
-            sorted(rs.GRANULARITIES)
-        )
 
     def test__test_granularities_constant(self):
         # Smoke test to catch myself when this changes.
